pyproject.toml
[project]
name = "version-v4"
version = "0.1.0"
description = "Fichiers de Poetry réparés"
authors = [
    {name = "Zoléni Kokolo Zassi",email = "zolenikokolozassi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce (>=2.5.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false